requirements.txt
attrs==21.4.0
black==22.3.0
click==8.1.3
flake8==4.0.1
iniconfig==1.1.1
mccabe
mypy-extensions==0.4.3
packaging==21.3
pathspec==0.9.0
platformdirs==2.5.2
pluggy==1.0.0
py==1.11.0
pycodestyle==2.8.0
pyflakes==2.4.0
pyparsing==3.0.9
pytest==7.1.2
tomli==2.0.1

Django==4.0.5
colorama~=0.4.4
pip~=22.1.2
Pillow~=9.1.1
docutils~=0.18.1
Sphinx~=5.0.0
tornado~=6.1
Jinja2~=3.1.2
pytz~=2022.1
sqlparse~=0.4.2
asgiref~=3.5.2
numpy~=1.22.4
wcwidth~=0.2.5
pexpect~=4.8.0
setuptools~=62.3.2
djangorestframework~=3.13.1
requests~=2.27.1
Pygments~=2.12.0
psycopg2-binary==2.9.3